[MDH-15] refactor : 웨이팅 패키지 구조 변경
