+++ src/main/java/com/mdh/devtable/waiting/domain/ShopWaitingStatus.java
@@ -1,4 +1,4 @@
-package com.mdh.devtable.waiting;
+package com.mdh.devtable.waiting.domain;
 
 import lombok.RequiredArgsConstructor;
